+++ DTROverlay/OverlayPositionOrigin.cs
@@ -4,4 +4,6 @@ public enum OverlayPositionOrigin
 {
     TopLeft,
     TopRight,
+    BottomLeft,
+    BottomRight,
 }